Assign data._staged after storing in assignee
The value of `data._staged` is lost after assigning to
`self.model.assignee.data`. This is likely because Pydantic is
configured to validate assignments, and `_staged` is a `PrivateAttr`.
The result is that underlying `data` object is copied by value and not
by reference, and `_staged` is initialized with it's default value of
`False.` You can validate this by printing `id(data)` and
`id(self.model.assignee.data)` after the assignment.

This was causing the staged assignee changes to not be propagated upon
calling `case.update()` as expected.